VER-91956: move test to a new location
